Revert pkg rename and provide vkey string constructor instead of note-compatible verifier.
